A little more code motion and cleanup
